move obstacle avoidance dialogs out of Exclusion Zone Panel
