+++ src/web/components/ObstacleAvoidanceDialogs/WaypointRemovalDialog/WaypointRemovalDialog.tsx
@@ -1,9 +1,9 @@
 import { useContext } from "react";
-import { JaiaContext, JaiaDispatchContext } from "../../../../context/JaiaContext";
-import { JaiaActions } from "../../../../context/jaia-actions";
-import { MAX_WAYPOINTS } from "../../../../utils/constants";
-import { ProposalStatus } from "../../../../data/obstacle_avoidance_data/pending-route-data";
-import ExclusionZoneBaseDialog from "../ExclusionZoneBaseDialog";
+import { JaiaContext, JaiaDispatchContext } from "../../../context/JaiaContext";
+import { JaiaActions } from "../../../context/jaia-actions";
+import { MAX_WAYPOINTS } from "../../../utils/constants";
+import { ProposalStatus } from "../../../data/obstacle_avoidance_data/pending-route-data";
+import ObstacleAvoidanceBaseDialog from "../ObstacleAvoidanceBaseDialog";
 import RerouteSummary from "../RerouteSummary";
 
 /**
@@ -42,7 +42,7 @@ export default function WaypointRemovalDialog() {
     }
 
     return (
-        <ExclusionZoneBaseDialog title="Mission Plan Update Required" buttons={buttons}>
+        <ObstacleAvoidanceBaseDialog title="Mission Plan Update Required" buttons={buttons}>
             <p>
                 <strong>{pending.totalRemovedCount}</strong> waypoint
                 {pending.totalRemovedCount !== 1 ? "s" : ""} inside an exclusion zone will be
@@ -77,6 +77,6 @@ export default function WaypointRemovalDialog() {
                 overLimitMessage={`The following mission${rerouteOverLimit.length !== 1 ? "s" : ""} still cross a zone after waypoint removal but cannot be rerouted — adding bypass waypoints would exceed the ${MAX_WAYPOINTS}-waypoint limit:`}
                 impossibleMessage={`The following mission${rerouteImpossible.length !== 1 ? "s" : ""} still cross a zone after waypoint removal and cannot be rerouted with the current zone layout:`}
             />
-        </ExclusionZoneBaseDialog>
+        </ObstacleAvoidanceBaseDialog>
     );
 }
