+++ src/web/components/ObstacleAvoidanceDialogs/ZoneCrossingDialog/ZoneCrossingDialog.tsx
@@ -1,4 +1,4 @@
-import ExclusionZoneBaseDialog from "../ExclusionZoneBaseDialog";
+import ObstacleAvoidanceBaseDialog from "../ObstacleAvoidanceBaseDialog";
 
 interface ZoneCrossingDialogProps {
     waypointNumber: number;
@@ -14,7 +14,7 @@ export default function ZoneCrossingDialog({
     onCancel,
 }: ZoneCrossingDialogProps) {
     return (
-        <ExclusionZoneBaseDialog
+        <ObstacleAvoidanceBaseDialog
             title="Route Crosses a Zone"
             buttons={[
                 { label: "Cancel", onClick: onCancel },
@@ -25,6 +25,6 @@ export default function ZoneCrossingDialog({
                 The path to waypoint {waypointNumber} has been rerouted to include{" "}
                 <strong>{bypassCount}</strong> bypass waypoint{bypassCount !== 1 ? "s" : ""}.
             </p>
-        </ExclusionZoneBaseDialog>
+        </ObstacleAvoidanceBaseDialog>
     );
 }
